solver-openmp: Deleted obsolete solvers.
According to [1], the predictor-corrector method and the Runge-Kutta
approach do not preserve the positivity of the density matrix in all
cases. Therefore, we drop the support of those solvers and focus on
the development of alternative methods that solve the Liouville-von
Neumann equation exactly for one time step.

This commit also deletes the last traces of Xeon Phi KNC support, which
was used by some of the deleted solvers.

[1] Riesch M. and Jirauschek C., Analyzing the positivity preservation
    of numerical methods for the Liouville-von Neumann equation,
    J. Comp. Phys. 390, 290--296