mbsolve-lib: Revised creational pattern for solvers and writers.
This commit thoroughly revises the way how instances of solvers and writers
are created. The revisions enable the build of static libraries,
which caused problems before. Additionally, based on the recommendation of
Tien Dat Nguyen the cpu-fdtd implementations are moved to *.tpp files in
order to allow fine-grained compilation. Thereby, the compilation may run
in parallel.