mbsolve-lib: Enabled Lindblad only simulation (0D Maxwell-Bloch).
This commit adjusts the Song 2005 setup so that it is treated as 0D Maxwell-
Bloch simulation (i.e., only one driven quantum mechanical system is
modeled by the Lindblad equation).

In the base library, mainly the time step calculation had to be adjusted
in order to enable this. The number of time steps can now be specified in
the scenario and is considered if the number of spatial grid points is 1.

Additionally, the time step calculation can now handle zero simulation end
time, in which case the number of time steps is set to 1.