mbsolve-lib: Added flexible boundary conditions for the EM field.
This commit introduces a framework to set the boundary conditions of a
device in a flexible way. Hereby, the user can specify semitransparent
mirror, perfectly reflecting mirror, and perfectly matched layer (PML)
boundary conditions by specifying certain reflectivity values.

Please note that support for the flexible boundary conditions must be
introduced in the solver implementations before they can be used in a
simulation.